# lib/api-zod/src/generated/types/authSignupBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthSignupRole } from './authSignupRole';

export interface AuthSignupBody {
  role: AuthSignupRole;
  /** @minLength 2 */
  fullName: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
}