# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthRole } from './authRole';

export interface AuthUser {
  id: number;
  fullName: string;
  email: string;
  role: AuthRole;
}