# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: number;
  title: string;
  category: string;
  oppositeParty: string;
  status: string;
  statusLabel: string;
  nextStep: string;
  updatedAt: string;
  progress: number;
  /** @nullable */
  lawyerName?: string | null;
}