# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';

export interface Dashboard {
  userName: string;
  openCases: number;
  upcomingAppointments: number;
  documents: number;
  unreadNotifications: number;
  recentActivity: Activity[];
}