# lib/api-zod/src/generated/types/law.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Law {
  id: number;
  title: string;
  category: string;
  summary: string;
  sections: string;
  updatedAt: string;
}