# lib/api-zod/src/generated/types/lawyer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lawyer {
  id: number;
  name: string;
  initials: string;
  specialization: string;
  experience: number;
  location: string;
  rating: number;
  reviews: number;
  fee: number;
  verified: boolean;
  availability: string;
}