# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: number;
  title: string;
  category: string;
  date: string;
  readTime: string;
  summary: string;
}