# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  description: string;
  date: string;
  amount: number;
  status: string;
  receipt: string;
}